requirements.txt
Pillow==7.0.0
Telethon==1.10.10
